# app/static/swagger.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: "Lunch API"
  description: "This is a Lunch API"
servers:
  - url: http://0.0.0.0:5000
...
paths:
  /api/restaurants/:
    get:
      tags:
        - "Restaurants"
      summary: "Get restaurants information"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestaurantsOut'
    post:
      security:
        - bearerAuth: []
      tags:
        - "Restaurants"
      summary: "Create new restaurant"
      description: "This can only be done by the logged in admin"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestaurantIn'
      responses:
        '201':
          description: "Successful Operation"
          content:
            application/json:
              example:
                id: "5"
        '400':
          description: "Wrong input data"
          content:
            application/json:
              example:
                message: "Please, specify 'name'"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
  /api/restaurants/{id}:
    get:
      tags:
        - "Restaurants"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of restaurant"
          required: true
          schema:
            type: "integer"
      summary: "Get restaurant by id"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestaurantOut'
        '404':
          description: "Restaurant not found"
          content:
            application/json:
              example:
                message: "Restaurant not found"
    patch:
      security:
        - bearerAuth: []
      tags:
        - "Restaurants"
      summary: "Update restaurant's information"
      description: "This can only be done by the logged in admin"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of restaurant"
          required: true
          schema:
            type: "integer"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestaurantIn'
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              example:
                message: "Updated"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
        '404':
          description: "Wrong restaurant's id"
          content:
            application/json:
              example:
                message: "Restaurant not found"
    delete:
      security:
        - bearerAuth: []
      tags:
        - "Restaurants"
      summary: "Delete restaurant"
      description: "This can only be done by the logged in admin"
      parameters:
        - name: "id"
          in: "path"
          description: "Id of the restaurant"
          required: true
          schema:
            type: "integer"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              example:
                message: "Deleted"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
        '404':
          description: "Wrong restaurant's id"
          content:
            application/json:
              example:
                message: "Restaurant not found"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RestaurantOut:
      type: "object"
      properties:
        id:
          type: "integer"
          example: 1
        name:
          type: "string"
          example: "McDonald's"
        menus:
          type: "array"
          example: []
    RestaurantsOut:
      type: "array"
      items:
        $ref: '#/components/schemas/RestaurantOut'
    RestaurantIn:
      type: "object"
      properties:
        name:
          type: "string"
          example: "Starbucks"
...